+++ console/src/pages/Settings/Debug/components/index.ts
@@ -1,0 +1,1 @@
+export { LogViewer } from "./LogViewer";
